refactor(types): separate hook and component types
