+++ apps/renderer/src/types/component-types.ts
@@ -56,19 +56,6 @@ export interface ToastProps {
   duration?: number;
 }
 
-export interface UseSearchProps {
-  query: string;
-  currentMatch: number;
-  matchCount: number;
-  isSearchOpen: boolean;
-  openSearch: () => void;
-  closeSearch: () => void;
-  setQuery: (q: string) => void;
-  goToNextMatch: () => void;
-  goToPrevMatch: () => void;
-  getHiglightedHtml: (html: string) => string;
-}
-
 export interface SearchBarProps {
   query: string;
   matchCount: number;
@@ -153,39 +140,19 @@ export type IconProps = React.SVGProps<SVGSVGElement> & {
   size?: number;
 };
 
-export interface UseMenuEventsProps {
-  onOpenFile: () => void;
-  onOpenFolder: () => void;
-  onSearchDocument: () => void;
-  onToggleToc: () => void;
-  onToggleBrowser: () => void;
-  onFocusMode: () => void;
-  onCycleTheme: () => void;
-  onZoomIn: () => void;
-  onZoomOut: () => void;
-  onZoomReset: () => void;
-  onNextTab: () => void;
-  onPreviousTab: () => void;
-  onCloseTab: () => void;
-  onExportHtml: () => void;
-  onExportPdf: () => void;
-  onExportDocx: () => void;
-}
-
-export interface UseShortcutsProps {
-  onOpenFile: () => void;
-  onOpenFolder: () => void;
-  onToggleFocusMode: () => void;
-  onToggleTheme: () => void;
-  onOpenSearch: () => void;
-  onCloseSearch: () => void;
-  onZoomIn: () => void;
-  onZoomOut: () => void;
-  onZoomReset: () => void;
-  onToggleSidebar: () => void;
-  onToggleFileBrowser: () => void;
-}
-
 export interface ElectronFile extends File {
   path: string;
+}
+
+export type ActiveTab = {
+  html: string;
+} | null;
+
+export interface ReaderToolbarProps {
+  fontSize: number;
+  theme: string;
+  onZoomIn: () => void;
+  onZoomOut: () => void;
+  onZoomReset: () => void;
+  onToggleTheme: () => void;
 }
